server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
